refactor(rng): use shared chacha8 rng helper
